refactor: implement small alert and replace warn banner
